+++ src/c/data.c
@@ -10,6 +10,22 @@
 #include <stdarg.h>
 #include <math.h>
 
+/* IOT Data Type      Key      Composed  Constant
+ *
+ * Integer Types      Y        N         N
+ * Boolean            Y        N         Y
+ * Pointer            Y        N         N
+ * String             Y        N         Can be
+ * Null               N        N         Y
+ * Binary             Y        N         N
+ * Array              Y        N         N
+ * Vector             Y        Y         N
+ * List               Y        Y         N
+ * Map                Y        Y         N
+*/
+
+#define IOT_DATA_IS_COMPOSED_TYPE(t) ((t) >= IOT_DATA_VECTOR && (t) <= IOT_DATA_MAP)
+
 #ifdef IOT_HAS_UUID
 #include <uuid/uuid.h>
 #ifndef UUID_STR_LEN
@@ -90,6 +106,8 @@ struct iot_data_t
   bool release_block : 1;
   bool heap : 1;
   bool constant : 1;
+  bool composed : 1;
+  bool rehash : 1;
 };
 
 typedef struct iot_data_value_base_t
@@ -231,7 +249,7 @@ static iot_data_value_base_t iot_data_null = { .base.type = IOT_DATA_NULL, .base
 extern void iot_data_init (void);
 extern void iot_data_map_dump (iot_data_t * map);
 static iot_data_t * iot_data_value_from_json (iot_json_tok_t ** tokens, const char * json, bool ordered, iot_data_t * cache);
-static void iot_node_free (iot_node_t * node);
+static void iot_node_free (iot_data_map_t * map, iot_node_t * node);
 static iot_node_t * iot_node_next (iot_node_t * iter);
 static iot_node_t * iot_node_prev (iot_node_t * iter);
 static bool iot_node_add (iot_data_map_t * map, iot_data_t * key, iot_data_t * value);
@@ -329,9 +347,50 @@ static inline void iot_data_block_free (iot_block_t * block)
 #endif
 }
 
+static inline void iot_data_map_hash (iot_data_t * map, const iot_data_t * key, const iot_data_t * value)
+{
+  uint32_t key_hash = iot_data_hash (key);
+  uint32_t val_hash = iot_data_hash (value);
+  map->hash ^= key_hash;
+  if (val_hash != key_hash) map->hash ^= val_hash; // Only apply value hash if different from key hash or hashes cancel out
+}
+
 uint32_t iot_data_hash (const iot_data_t * data)
 {
-  return data ? data->hash : 0u;
+  iot_data_t * da = (iot_data_t*) data;
+  if (da && da->rehash)
+  {
+    da->hash = 0;
+    da->rehash = false;
+    if (da->type == IOT_DATA_VECTOR)
+    {
+      iot_data_vector_iter_t iter;
+      iot_data_vector_iter (da, &iter);
+      while (iot_data_vector_iter_next (&iter))
+      {
+        da->hash ^= iot_data_hash (iot_data_vector_iter_value (&iter));
+      }
+    }
+    else if (da->type == IOT_DATA_LIST)
+    {
+      iot_data_list_iter_t iter;
+      iot_data_list_iter (da, &iter);
+      while (iot_data_list_iter_next (&iter))
+      {
+        da->hash ^= iot_data_hash (iot_data_list_iter_value (&iter));
+      }
+    }
+    else // IOT_DATA_MAP
+    {
+      iot_data_map_iter_t iter;
+      iot_data_map_iter (da, &iter);
+      while (iot_data_map_iter_next (&iter))
+      {
+        iot_data_map_hash (da, iot_data_map_iter_key (&iter), iot_data_map_iter_value (&iter));
+      }
+    }
+  }
+  return da ? da->hash : 0u;
 }
 
 static inline void iot_element_free (iot_element_t * element)
@@ -365,6 +424,7 @@ static void * iot_data_block_alloc_data (iot_data_type_t type)
   bool heap = iot_data_alloc_from_heap;
   iot_data_t * data = heap ? calloc (1, IOT_DATA_BLOCK_SIZE) : iot_data_block_alloc ();
   data->heap = heap;
+  data->composed = IOT_DATA_IS_COMPOSED_TYPE (type);
   iot_data_block_init (data, type);
   return data;
 }
@@ -427,6 +487,11 @@ iot_data_t * iot_data_add_ref (const iot_data_t * data)
   return (iot_data_t*) data;
 }
 
+uint32_t iot_data_ref_count (const iot_data_t * data)
+{
+  return data ? atomic_load (&data->refs) : 0;
+}
+
 iot_data_type_t iot_data_name_type (const char * name)
 {
   iot_data_type_t type = 0;
@@ -462,24 +527,15 @@ extern const iot_data_t * iot_data_get_metadata (const iot_data_t * data)
   return data ? data->base.meta : NULL;
 }
 
-static int iot_data_key_cmp (const iot_data_t * v1, const iot_data_t * v2)
+static int iot_data_compare (const iot_data_t * v1, const iot_data_t * v2)
 {
   if (v1->type == v2->type)
   {
-    switch (v1->type)
+    if (v1->type == IOT_DATA_STRING)
     {
-      case IOT_DATA_STRING:
-        return strcmp (((const iot_data_value_t *) v1)->value.str, ((const iot_data_value_t *) v2)->value.str);
-      case IOT_DATA_BINARY:
-      case IOT_DATA_ARRAY:
-        return (iot_data_equal (v1, v2)) ? 0 : ((v1->hash < v2->hash) ? -1 : 1);
-      default:
-      {
-        uint64_t ui1 = ((const iot_data_value_t *) v1)->value.ui64;
-        uint64_t ui2 = ((const iot_data_value_t *) v2)->value.ui64;
-        return (ui1 == ui2) ? 0 : ((ui1 < ui2) ? -1 : 1);
-      }
+      return strcmp (((const iot_data_value_t *) v1)->value.str, ((const iot_data_value_t *) v2)->value.str);
     }
+    return (iot_data_equal (v1, v2)) ? 0 : ((iot_data_hash (v1) < iot_data_hash (v2)) ? -1 : 1);
   }
   return (v1->type < v2->type) ? -1 : 1;
 }
@@ -554,7 +610,7 @@ bool iot_data_equal (const iot_data_t * v1, const iot_data_t * v2)
 
 iot_data_t * iot_data_alloc_map (iot_data_type_t key_type)
 {
-  assert (key_type < IOT_DATA_MAP && key_type != IOT_DATA_NULL);
+  assert (key_type != IOT_DATA_NULL);
   iot_data_map_t * map = iot_data_block_alloc_data (IOT_DATA_MAP);
   map->base.key_type = key_type;
   return (iot_data_t*) map;
@@ -667,6 +723,7 @@ extern bool iot_data_list_remove (iot_data_t * list, iot_data_cmp_fn cmp, const 
       element->prev->next = element->next;
       impl->head->length--;
       iot_element_free (element);
+      list->rehash = true;
     }
     iot_data_free (value);
   }
@@ -722,7 +779,12 @@ iot_data_t * iot_data_list_iter_replace (const iot_data_list_iter_t * iter, iot_
 {
   assert (iter && iter->_list && value && (iter->_list->base.element_type == IOT_DATA_MULTI || iter->_list->base.element_type == value->type));
   iot_data_t * res = (iter->_element) ? iter->_element->value : NULL;
-  if (res) iter->_element->value = value;
+  if (res)
+  {
+    iot_data_t * base = (iot_data_t*) &iter->_list->base;
+    base->rehash = true;
+    iter->_element->value = value;
+  }
   return res;
 }
 
@@ -743,6 +805,7 @@ void iot_data_list_tail_push (iot_data_t * list, iot_data_t * value)
     impl->head = impl->tail;
   }
   impl->head->length++;
+  list->hash ^= iot_data_hash (value);
 }
 
 iot_data_t * iot_data_list_tail_pop (iot_data_t * list)
@@ -764,6 +827,7 @@ iot_data_t * iot_data_list_tail_pop (iot_data_t * list)
     {
       impl->head = NULL;
     }
+    list->rehash = true;
     iot_element_free (element);
   }
   return value;
@@ -787,6 +851,7 @@ void iot_data_list_head_push (iot_data_t * list, iot_data_t * value)
     impl->tail = impl->head;
     impl->head->length = 1;
   }
+  list->hash ^= iot_data_hash (value);
 }
 
 iot_data_t * iot_data_list_head_pop (iot_data_t * list)
@@ -808,6 +873,7 @@ iot_data_t * iot_data_list_head_pop (iot_data_t * list)
     {
       impl->tail = NULL;
     }
+    list->rehash = true;
     iot_element_free (element);
   }
   return value;
@@ -827,6 +893,68 @@ bool iot_data_is_of_type (const iot_data_t * data, iot_data_type_t type)
 bool iot_data_is_static (const iot_data_t * data)
 {
   return (data && data->constant);
+}
+
+static void iot_data_cache_add (iot_data_t * cache, iot_data_t ** data)
+{
+  iot_data_t * cached = (iot_data_t*) iot_data_map_get (cache, *data);
+  if (cached)
+  {
+    iot_data_free (*data);
+    iot_data_add_ref (cached);
+    *data = cached;
+  }
+  else
+  {
+    iot_data_map_add (cache, iot_data_add_ref (*data), iot_data_add_ref (*data));
+    if ((*data)->composed)
+    {
+      iot_data_compress_with_cache (*data, cache);
+    }
+  }
+}
+
+void iot_data_compress_with_cache (iot_data_t * data, iot_data_t * cache)
+{
+  assert (data && cache && (cache->key_type == IOT_DATA_MULTI) && (cache->element_type == IOT_DATA_MULTI));
+  if (data->composed)
+  {
+    if (data->type == IOT_DATA_VECTOR)
+    {
+      iot_data_vector_iter_t iter;
+      iot_data_vector_iter (data, &iter);
+      while (iot_data_vector_iter_next (&iter))
+      {
+        iot_data_cache_add (cache, &(iter._vector->values[iter._index]));
+      }
+    }
+    else if (data->type == IOT_DATA_LIST)
+    {
+      iot_data_list_iter_t iter;
+      iot_data_list_iter (data, &iter);
+      while (iot_data_list_iter_next (&iter))
+      {
+        iot_data_cache_add (cache, &(iter._element->value));
+      }
+    }
+    else // Map
+    {
+      iot_data_map_iter_t iter;
+      iot_data_map_iter (data, &iter);
+      while (iot_data_map_iter_next (&iter))
+      {
+        iot_data_cache_add (cache, &(iter._node->value));
+        iot_data_cache_add (cache, &(iter._node->key));
+      }
+    }
+  }
+}
+
+void iot_data_compress (iot_data_t * data)
+{
+  iot_data_t * cache = iot_data_alloc_typed_map (IOT_DATA_MULTI, IOT_DATA_MULTI);
+  iot_data_compress_with_cache (data, cache);
+  iot_data_free (cache);
 }
 
 const void * iot_data_address (const iot_data_t * data)
@@ -864,7 +992,7 @@ void iot_data_free (iot_data_t * data)
       }
       case IOT_DATA_MAP:
       {
-        iot_node_free (((iot_data_map_t*) data)->tree);
+        iot_node_free ((iot_data_map_t*) data, ((iot_data_map_t*) data)->tree);
         break;
       }
       case IOT_DATA_VECTOR:
@@ -962,6 +1090,7 @@ iot_data_t * iot_data_alloc_i8 (int8_t val)
 {
   iot_data_value_t * data = iot_data_value_alloc (IOT_DATA_INT8, false);
   data->value.i8 = val;
+  data->base.hash = (uint32_t) val;
   return (iot_data_t*) data;
 }
 
@@ -969,6 +1098,7 @@ iot_data_t * iot_data_alloc_ui8 (uint8_t val)
 {
   iot_data_value_t * data = iot_data_value_alloc (IOT_DATA_UINT8, false);
   data->value.ui8 = val;
+  data->base.hash = (uint32_t) val;
   return (iot_data_t*) data;
 }
 
@@ -976,6 +1106,7 @@ iot_data_t * iot_data_alloc_i16 (int16_t val)
 {
   iot_data_value_t * data = iot_data_value_alloc (IOT_DATA_INT16, false);
   data->value.i16 = val;
+  data->base.hash = (uint32_t) val;
   return (iot_data_t*) data;
 }
 
@@ -983,6 +1114,7 @@ iot_data_t * iot_data_alloc_ui16 (uint16_t val)
 {
   iot_data_value_t * data = iot_data_value_alloc (IOT_DATA_UINT16, false);
   data->value.ui16 = val;
+  data->base.hash = (uint32_t) val;
   return (iot_data_t*) data;
 }
 
@@ -990,6 +1122,7 @@ iot_data_t * iot_data_alloc_i32 (int32_t val)
 {
   iot_data_value_t * data = iot_data_value_alloc (IOT_DATA_INT32, false);
   data->value.i32 = val;
+  data->base.hash = (uint32_t) val;
   return (iot_data_t*) data;
 }
 
@@ -997,6 +1130,7 @@ iot_data_t * iot_data_alloc_ui32 (uint32_t val)
 {
   iot_data_value_t * data = iot_data_value_alloc (IOT_DATA_UINT32, false);
   data->value.ui32 = val;
+  data->base.hash = val;
   return (iot_data_t*) data;
 }
 
@@ -1004,6 +1138,7 @@ iot_data_t * iot_data_alloc_i64 (int64_t val)
 {
   iot_data_value_t * data = iot_data_value_alloc (IOT_DATA_INT64, false);
   data->value.i64 = val;
+  data->base.hash = (uint32_t) val;
   return (iot_data_t*) data;
 }
 
@@ -1011,6 +1146,7 @@ iot_data_t * iot_data_alloc_ui64 (uint64_t val)
 {
   iot_data_value_t * data = iot_data_value_alloc (IOT_DATA_UINT64, false);
   data->value.ui64 = val;
+  data->base.hash = (uint32_t) val;
   return (iot_data_t*) data;
 }
 
@@ -1018,6 +1154,7 @@ iot_data_t * iot_data_alloc_f32 (float val)
 {
   iot_data_value_t * data = iot_data_value_alloc (IOT_DATA_FLOAT32, false);
   data->value.f32 = val;
+  data->base.hash = (uint32_t) val;
   return (iot_data_t*) data;
 }
 
@@ -1025,6 +1162,7 @@ iot_data_t * iot_data_alloc_f64 (double val)
 {
   iot_data_value_t * data = iot_data_value_alloc (IOT_DATA_FLOAT64, false);
   data->value.f64 = val;
+  data->base.hash = (uint32_t) val;
   return (iot_data_t*) data;
 }
 
@@ -1302,6 +1440,7 @@ bool iot_data_map_remove (iot_data_t * map, const iot_data_t * key)
     iot_data_map_t * mp = (iot_data_map_t*) map;
     if ((ret = iot_node_remove (mp, key)))
     {
+      mp->base.rehash = true;
       mp->size--;
     }
   }
@@ -1328,9 +1467,10 @@ bool iot_data_string_map_remove (iot_data_t * map, const char * key)
 
 void iot_data_map_add (iot_data_t * map, iot_data_t * key, iot_data_t * val)
 {
-  assert (map && (map->type == IOT_DATA_MAP));
-  assert (key && key->type == map->key_type);
-  assert (val && (map->element_type == IOT_DATA_MULTI || map->element_type == val->type)); // Check value type matches for fixed type map
+  assert (map && key && val);
+  assert (map->type == IOT_DATA_MAP && key->type != IOT_DATA_NULL);
+  assert (key->type == map->key_type || map->key_type == IOT_DATA_MULTI);
+  assert (map->element_type == val->type || map->element_type == IOT_DATA_MULTI);
   if (iot_node_add ((iot_data_map_t*) map, key, val))
   {
     ((iot_data_map_t*) map)->size++;
@@ -1492,9 +1632,17 @@ void iot_data_vector_add (iot_data_t * vector, uint32_t index, iot_data_t * val)
   iot_data_vector_t * arr = (iot_data_vector_t*) vector;
   assert (vector && (vector->type == IOT_DATA_VECTOR));
   assert (index < arr->size);
-  assert (val && (vector->element_type == IOT_DATA_MULTI || vector->element_type == val->type)); // Check element type matches for fixed type vector
+  assert ((val == NULL) || (val && (vector->element_type == IOT_DATA_MULTI || vector->element_type == val->type))); // Check element type matches for fixed type vector
   iot_data_t * element = arr->values[index];
-  iot_data_free (element);
+  if (element)
+  {
+    vector->rehash = true;
+    iot_data_free (element);
+  }
+  else
+  {
+    vector->hash ^= iot_data_hash (val);
+  }
   arr->values[index] = val;
 }
 
@@ -1706,10 +1854,19 @@ iot_data_t * iot_data_vector_iter_replace_value (const iot_data_vector_iter_t * 
 {
   assert (iter && iter->_vector && value && (iter->_vector->base.element_type == IOT_DATA_MULTI || iter->_vector->base.element_type == value->type));
   iot_data_t * res = NULL;
-  if (iter->_index < iter->_vector->size)
+  iot_data_vector_t * vector = (iot_data_vector_t*) iter->_vector;
+  if (iter->_index < vector->size)
   {
-    res = iter->_vector->values[iter->_index];
-    iter->_vector->values[iter->_index] = value;
+    res = vector->values[iter->_index];
+    if (res)
+    {
+      vector->base.rehash = true;
+    }
+    else
+    {
+      if (value) vector->base.hash ^= iot_data_hash (value);
+    }
+    vector->values[iter->_index] = value;
   }
   return res;
 }
@@ -2615,7 +2772,7 @@ static inline iot_node_t * iot_node_find (const iot_node_t * node, const iot_dat
 {
   while (node)
   {
-    int cmp = iot_data_key_cmp (node->key, key);
+    int cmp = iot_data_compare (node->key, key);
     if (cmp == 0) break;
     node = (cmp > 0) ? node->left : node->right;
   }
@@ -2630,11 +2787,11 @@ static void iot_node_insert (iot_data_map_t * map, iot_data_t * key, iot_data_t 
   while (x)
   {
     y = x;
-    x = (iot_data_key_cmp (key, x->key) < 0) ? x->left : x->right;
+    x = (iot_data_compare (key, x->key) < 0) ? x->left : x->right;
   }
   node->parent = y;
   if (y == NULL) map->tree = node;
-  else if (iot_data_key_cmp (key, y->key) < 0) y->left = node;
+  else if (iot_data_compare (key, y->key) < 0) y->left = node;
   else y->right = node;
 
   if (node->parent)
@@ -2704,23 +2861,26 @@ static bool iot_node_add (iot_data_map_t * map, iot_data_t * key, iot_data_t * v
   iot_node_t * node = iot_node_find (map->tree, key);
   if (node)
   {
+    map->base.rehash = true;
     iot_data_free (key);
     iot_data_free (node->value);
     node->value = value;
+    key = node->key;
   }
   else
   {
     iot_node_insert (map, key, value);
+    iot_data_map_hash (&map->base, key, value);
   }
   return (node == NULL);
 }
 
-static void iot_node_free (iot_node_t * node)
+static void iot_node_free (iot_data_map_t * map, iot_node_t * node)
 {
   if (node)
   {
-    iot_node_free (node->left);
-    iot_node_free (node->right);
+    iot_node_free (map, node->left);
+    iot_node_free (map, node->right);
     iot_node_delete (node);
   }
 }