+++ RELEASE_NOTES.md
@@ -208,6 +208,12 @@
 
 * `iot_data_hash`
 
+- Functions added to compress composed data values (Map, Vector and List):
+
+* `iot_data_compress`
+* `iot_data_compress_with_cache`
+
+- Map data type updated to support all key types
 - Fedora support updated for Fedora 35
 - Support removed for Alpine Linux versions 3.12 and 3.13
 - Support removed for CentOS 8